requirements.txt
opencv-python
deepface
python-dotenv
spotipy
argparse